Refactor: Move OCF exit codes to results.h
